% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\name{install_nwb}
\alias{install_nwb}
\title{Install \code{'NWB'} via \code{'pynwb'}}
\usage{
install_nwb(python_ver = "auto", verbose = TRUE)
}
\arguments{
\item{python_ver}{'Python' version, see \code{\link[rpymat]{configure_conda}};
default is \code{"auto"}, which is suggested}

\item{verbose}{whether to print the installation messages}
}
\value{
This function returns nothing.
}
\description{
Install \code{'NWB'} via \code{'pynwb'}
}
