% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-NWBHDF5IO.R
\name{NWBHDF5IO}
\alias{NWBHDF5IO}
\title{Creates a \code{NWBHDF5IO} file container}
\value{
Nothing

Nothing

\code{'NWBFile'} container

Whatever results generated by \code{expr}
}
\description{
Creates a \code{NWBHDF5IO} file container

Creates a \code{NWBHDF5IO} file container
}
\examples{
\dontrun{

# Running this example requires a .nwb file

library(rnwb)
container <- NWBHDF5IO$new(path = file)
container$with({

  data <- container$read()
  electrode_table <- data$electrodes[convert = TRUE]

})

print(electrode_table)

}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{opened}}{Whether the container is opened.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-NWBHDF5IO-new}{\code{NWBHDF5IO$new()}}
\item \href{#method-NWBHDF5IO-close}{\code{NWBHDF5IO$close()}}
\item \href{#method-NWBHDF5IO-close_linked_files}{\code{NWBHDF5IO$close_linked_files()}}
\item \href{#method-NWBHDF5IO-read}{\code{NWBHDF5IO$read()}}
\item \href{#method-NWBHDF5IO-with}{\code{NWBHDF5IO$with()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NWBHDF5IO-new"></a>}}
\if{latex}{\out{\hypertarget{method-NWBHDF5IO-new}{}}}
\subsection{Method \code{new()}}{
Initialize the class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NWBHDF5IO$new(path = NULL, mode = c("r", "w", "r+", "a", "w-", "x"), ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{path}}{Path to a \code{'.nwb'} file}

\item{\code{mode}}{Mode for opening the file}

\item{\code{...}}{Other parameters passed to \code{nwb$NWBHDF5IO}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NWBHDF5IO-close"></a>}}
\if{latex}{\out{\hypertarget{method-NWBHDF5IO-close}{}}}
\subsection{Method \code{close()}}{
Close the connections (low-level method, see 'with' method below)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NWBHDF5IO$close(close_links = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{close_links}}{Whether to close all files linked to from this
file; default is true}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NWBHDF5IO-close_linked_files"></a>}}
\if{latex}{\out{\hypertarget{method-NWBHDF5IO-close_linked_files}{}}}
\subsection{Method \code{close_linked_files()}}{
Close all opened, linked-to files. \code{'MacOS'} and \code{'Linux'}
automatically release the linked-to file after the linking file is
closed, but \code{'Windows'} does not, which prevents the linked-to
file from being deleted or truncated. Use this method to close all
opened, linked-to files.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NWBHDF5IO$close_linked_files()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NWBHDF5IO-read"></a>}}
\if{latex}{\out{\hypertarget{method-NWBHDF5IO-read}{}}}
\subsection{Method \code{read()}}{
Read the \code{'NWB'} file from the 'IO' source. Please use along with
\code{'$with'} method
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NWBHDF5IO$read()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-NWBHDF5IO-with"></a>}}
\if{latex}{\out{\hypertarget{method-NWBHDF5IO-with}{}}}
\subsection{Method \code{with()}}{
Safe wrapper for reading and handling \code{'NWB'} file. See class examples.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{NWBHDF5IO$with(expr, quoted = FALSE, envir = parent.frame())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expr}}{R expression to evaluate}

\item{\code{quoted}}{Whether \code{expr} is quoted; default is false}

\item{\code{envir}}{environment for \code{expr} to evaluate; default is the
parent frame (see \code{parent.frame})}
}
\if{html}{\out{</div>}}
}
}
}
